# apps/docs/src/openapi/tradr-api.json
{
...
  "info": {
    "description": "HTTP API for Tradr — the open-source trading journal with an AI advisor.\n\nThis reference is generated from the `@swagger` JSDoc blocks that live next to each route in the API source, so it always matches the endpoints the app actually serves. All routes are mounted under `/api`. The same API backs both the hosted app (`api.tradr.cloud`) and a self-hosted instance (reached through the web container at `/api`). Most endpoints require an authenticated session cookie.",
    "title": "Tradr API",
    "version": "0.5.3"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/accounts": {
      "post": {
        "description": "Authed. Creates a trading account for the current user. When feature gating is enabled and the user is a non-admin Free user at the L1 account cap, the create is refused with `403 TIER_LIMIT_ACCOUNTS` (plan-tiers REQ-6.1) — admins and gating-off deployments pass through unchanged.\n",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "brokerageId": {
                    "format": "uuid",
                    "nullable": true,
                    "type": "string"
                  },
                  "currency": {
                    "maxLength": 3,
                    "minLength": 3,
                    "type": "string"
                  },
                  "name": {
                    "minLength": 1,
                    "type": "string"
                  },
                  "startingBalance": {
                    "type": "string"
                  },
                  "timezone": {
                    "description": "IANA zone name (canonical spelling, e.g. `America/New_York`, `Etc/UTC`). Defines the account's trading day. Omitted defaults to `America/New_York`; an unknown zone is a 400.\n",
                    "example": "America/New_York",
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "currency"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The created account."
          },
          "400": {
            "description": "Validation error."
          },
          "403": {
            "description": "TIER_LIMIT_ACCOUNTS (account cap reached on the current plan) or FORBIDDEN (cross-user brokerage)."
          },
          "409": {
            "description": "Duplicate account name."
          }
        },
        "summary": "Create an account.",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/accounts/writable": {
      "put": {
        "description": "Authed, always-on. Stores which of the user's accounts stays writable for new trading data while over the Free-tier account cap. A plain stored preference — independent of gating/tier/over-cap state; it only takes effect when L1-writability enforcement consults the resolver. The account must belong to the current user (404 otherwise).\n",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "accountId": {
                    "format": "uuid",
                    "type": "string"
                  }
                },
                "required": [
                  "accountId"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "{ writableAccountId } — the stored designation."
          },
          "400": {
            "description": "Validation error."
          },
          "404": {
            "description": "Account not found (or not owned by the user)."
          }
        },
        "summary": "Set the writable-account designation.",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/accounts/{id}": {
      "put": {
        "description": "Authed. All fields optional. `startingBalance` is deliberately absent — it is creation-only. `timezone` IS editable: it changes only subsequent trading-day evaluations and rewrites no history.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "brokerageId": {
                    "format": "uuid",
                    "nullable": true,
                    "type": "string"
                  },
                  "currency": {
                    "maxLength": 3,
                    "minLength": 3,
                    "type": "string"
                  },
                  "name": {
                    "minLength": 1,
                    "type": "string"
                  },
                  "timezone": {
                    "description": "IANA zone name (canonical spelling). Unknown zone is a 400.",
                    "example": "America/New_York",
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated account."
          },
          "400": {
            "description": "Validation error (includes an unknown IANA timezone)."
          },
          "403": {
            "description": "FORBIDDEN (cross-user brokerage)."
          },
          "404": {
            "description": "Account not found."
          },
          "409": {
            "description": "Duplicate account name",
            "or currency change while positions exist.": null
          }
        },
        "summary": "Update an account.",
        "tags": [
          "Accounts"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Hosted (tradr.cloud)",
      "url": "https://api.tradr.cloud"
    },
    {
      "description": "Local development (API PORT default)",
      "url": "http://localhost:3100"
    }
  ],
...
}